pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.84.0",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.2.2",
    "pillow>=11.2.1",
    "jinja2>=3.1.6",
    "google-auth>=2.40.2",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "reportlab>=4.4.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
